Fix pyo3 pymethods
